requirements.txt
# Kaoruko Userbot Requirements

# Core
pyrogram>=2.0.106
tgcrypto>=1.2.5

# Database
motor>=3.3.2
pymongo>=4.6.1

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.1
aiohttp>=3.9.1
pillow>=10.1.0
requests>=2.31.0

# Optional but recommended
uvloop>=0.19.0  # For better async performance on Linux
